Document and licensify code
