+++ APIError.js
@@ -1,0 +1,23 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.*
+ */
+
+const winston = require('winston');
+
+/**
+ * A simple helper class for sending JSON-formatted errors to clients
+ */
+class APIError extends Error {
+  constructor(message) {
+    super(message);
+    winston.error('new API Error: ' + message);
+  }
+
+  toString() {
+    return JSON.stringify({error: true, message: this.message});
+  }
+}
+
+module.exports = APIError;
