+++ Engine.js
@@ -1,5 +1,14 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.*
+ */
+
 const winston = require('winston');
 
+/**
+ * An engine for running and managing list of rules
+ */
 class Engine {
   constructor() {
     this.rules = {};
@@ -53,6 +62,9 @@ class Engine {
     delete this.rules[ruleId];
   }
 
+  /**
+   * Update all active rules
+   */
   update() {
     winston.info('engine update');
     this.getRules().forEach(rule => {
