+++ Property.js
@@ -1,9 +1,22 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.*
+ */
+
 const assert = require('assert');
 const config = require('config');
 const fetch = require('node-fetch');
 const winston = require('winston');
 
+/**
+ * Utility to support operations on Thing's properties
+ */
 class Property {
+  /**
+   * Create a Property from a descriptor returned by the WoT API
+   * @param {PropertyDescription} desc
+   */
   constructor(desc) {
     assert(desc.type);
     assert(desc.href);
@@ -20,10 +33,16 @@ class Property {
     this.name = parts[parts.length - 1];
   }
 
+  /**
+   * @return {String} full property href
+   */
   getHref() {
     return config.get('gateway') + this.href + '?jwt=' + config.get('jwt');
   }
 
+  /**
+   * @return {Promise} resolves to property's value
+   */
   get() {
     winston.info('property got', {name: this.name});
     return fetch(this.getHref()).then(res => {
@@ -34,6 +53,10 @@ class Property {
     });
   }
 
+  /**
+   * @param {any} value
+   * @return {Promise} resolves if property is set to value
+   */
   set(value) {
     let data = {};
     data[this.name] = value;
