+++ actions/SetAction.js
@@ -1,6 +1,16 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.*
+ */
+
 const assert = require('assert');
 const Action = require('./Action');
 
+/**
+ * An Action which permanently sets the target property to
+ * a value when triggered
+ */
 class SetAction extends Action {
   constructor(desc) {
     super(desc);
@@ -10,6 +20,9 @@ class SetAction extends Action {
     this.on = false;
   }
 
+  /**
+   * @return {State}
+   */
   setState(state) {
     if (!this.on && state.on) {
       this.on = true;
