+++ static/index.html
@@ -25,7 +25,6 @@
       <span class="select select-action-value"></span>
       <input type="number" class="select select-action-value-input hidden" placeholder="0.0" step="0.1"/>
     </div>
-    <!-- <a href="#" id="create-rule">Create Rule</a> -->
     <script src="src/jwt.js"></script>
     <script src="src/Gateway.js"></script>
     <script src="src/Selector.js"></script>
