+++ triggers/BooleanTrigger.js
@@ -1,6 +1,16 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.*
+ */
+
 const assert = require('assert');
 const PropertyTrigger = require('./PropertyTrigger');
 
+/**
+ * A Trigger which activates when a boolean property is
+ * equal to a given value, `onValue`
+ */
 class BooleanTrigger extends PropertyTrigger {
   constructor(desc) {
     super(desc);
@@ -9,6 +19,9 @@ class BooleanTrigger extends PropertyTrigger {
     this.onValue = desc.onValue;
   }
 
+  /**
+   * @return {State}
+   */
   getState() {
     return this.property.get().then(propValue => {
       if (propValue === this.onValue) {
