+++ triggers/index.js
@@ -1,3 +1,9 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.*
+ */
+
 const triggers = {
   BooleanTrigger: require('./BooleanTrigger'),
   LevelTrigger: require('./LevelTrigger'),
@@ -5,6 +11,12 @@ const triggers = {
   Trigger: require('./Trigger')
 };
 
+/**
+ * Produce an trigger from a serialized trigger description. Throws if `desc`
+ * is invalid
+ * @param {TriggerDescription} desc
+ * @return {Trigger}
+ */
 function fromDescription(desc) {
   let TriggerClass = triggers[desc.type];
   if (!TriggerClass) {
